Updated with the review suggestions
