+++ src/main/java/com/avioconsulting/mule/logger/api/processor/FlowLogConfig.java
@@ -1,19 +1,43 @@
 package com.avioconsulting.mule.logger.api.processor;
 
 import org.mule.runtime.extension.api.annotation.Alias;
+import org.mule.runtime.extension.api.annotation.param.Optional;
 import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.ParameterGroup;
 import org.mule.runtime.extension.api.annotation.param.display.Summary;
 
-@Alias("flow-logs-config")
+@Alias("flow-log-config")
 public class FlowLogConfig {
 
   @Parameter
   @Summary("Name of the flow to associate given expression as attributes")
   private String flowName;
 
-  @ParameterGroup(name = "Flow Attributes")
-  private ExpressionText expressionText;
+  @Parameter
+  @Optional
+  @Summary("A valid dataweave expression that resolves to a Map object with key-value pairs")
+  private String attributesExpressionText;
+
+  @Parameter
+  @Optional
+  @Summary("A valid dataweave expression that results in a String to append to default flow start message")
+  private String messageExpressionText;
+
+  public String getAttributesExpressionText() {
+    return attributesExpressionText;
+  }
+
+  public void setAttributesExpressionText(String attributesExpressionText) {
+    this.attributesExpressionText = attributesExpressionText;
+  }
+
+  public String getMessageExpressionText() {
+    return messageExpressionText;
+  }
+
+  public void setMessageExpressionText(String messageExpressionText) {
+    this.messageExpressionText = messageExpressionText;
+  }
 
   public String getFlowName() {
     return flowName;
@@ -24,11 +48,4 @@ public class FlowLogConfig {
     return this;
   }
 
-  public ExpressionText getExpressionText() {
-    return expressionText;
-  }
-
-  public void setExpressionText(ExpressionText expressionText) {
-    this.expressionText = expressionText;
-  }
 }
