+++ src/test/resources/custom-logger-config.xml
@@ -9,14 +9,14 @@ http://www.mulesoft.org/schema/mule/avio-logger http://www.mulesoft.org/schema/m
     <avio-logger:config name="AVIO_Logger_Config_Json" doc:name="AVIO Logger Config" doc:id="84bd2712-015b-46b8-93e9-94536774b8ad"
                         applicationVersion="1" applicationName="munit" environment="test" compressor="GZIP" encryptionAlgorithm="PBEWithHmacSHA512AndAES_128" encryptionPassword="example" enableFlowLogs="true" formatAsJson="true" flowLogLevel="INFO">
         <avio-logger:flow-log-configs >
-            <avio-logger:flow-logs-config flowName="custom-logger-configFlow" attributesExpressionText="attributes" />
+            <avio-logger:flow-log-config flowName="custom-logger-configFlow" attributesExpressionText="attributes" />
         </avio-logger:flow-log-configs>
     </avio-logger:config>
 
     <avio-logger:config name="AVIO_Logger_Config" doc:name="AVIO Logger Config" doc:id="84bd2712-015b-46b8-93e9-94536774b8ad"
                       applicationVersion="1" applicationName="munit" environment="test" compressor="GZIP" encryptionAlgorithm="PBEWithHmacSHA512AndAES_128" encryptionPassword="example" enableFlowLogs="true" flowLogLevel="INFO">
         <avio-logger:flow-log-configs >
-            <avio-logger:flow-logs-config flowName="custom-logger-configFlow" attributesExpressionText="attributes" />
+            <avio-logger:flow-log-config flowName="custom-logger-configFlow" attributesExpressionText="attributes" />
         </avio-logger:flow-log-configs>
     </avio-logger:config>
